# static/openapi.yaml
openapi: 3.0.0
info:
  title: Cinemas and Movies API
  description: API to get information about cinemas and movies in Madrid.
  version: 1.0.0
servers:
  - url: http://localhost:5000
paths:
  /cinemas:
    get:
      summary: Get list of cinemas
      responses:
        '200':
          description: List of cinemas
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "E0146"
                    name:
                      type: string
                      example: "Cines Callao"
                    address:
                      type: string
                      example: "Plaza Callao 3 28013 Madrid"
                    num_screens:
                      type: string
                      example: "5"
                    url:
                      type: string
                      example: "https://www.sensacine.com/cines/cine/E0146/"
        '404':
          description: No cinema data available for today
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "No cinema data available for today"
  /movies:
    get:
      summary: Get list of movies in a cinema
      parameters:
        - in: query
          name: cinema_id
          schema:
            type: string
          required: true
          description: Cinema ID
      responses:
        '200':
          description: List of movies in the specified cinema
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "12345"
                    title:
                      type: string
                      example: "La virgen roja"
                    director:
                      type: string
                      example: "Example Director"
                    cast:
                      type: array
                      items:
                        type: string
                        example: "Example Actor"
                    synopsis:
                      type: string
                      example: "An example synopsis."
                    showtimes:
                      type: object
                      additionalProperties:
                        type: array
                        items:
                          type: string
                          example: "16:00"
        '404':
          description: Cinema not found or no data available
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Cinema not found"
  /showtimes:
    get:
      summary: Get showtimes of a movie in a specific cinema
      parameters:
        - in: query
          name: cinema_id
          schema:
            type: string
          required: true
          description: Cinema ID
        - in: query
          name: movie_title
          schema:
            type: string
          required: true
          description: Movie title
      responses:
        '200':
          description: Showtimes of the movie in the cinema
          content:
            application/json:
              schema:
                type: object
                properties:
                  cinema:
                    type: string
                    example: "Cines Callao"
                  movie:
                    type: string
                    example: "La virgen roja"
                  showtimes:
                    type: object
                    additionalProperties:
                      type: array
                      items:
                        type: string
                        example: "16:00"
        '404':
          description: Showtimes not found for the specified movie in the cinema
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Showtimes not found for the specified movie in the cinema"
  /cinemas_movie:
    get:
      summary: Get list of cinemas showing a specific movie
      parameters:
        - in: query
          name: movie_title
          schema:
            type: string
          required: true
          description: Movie title
      responses:
        '200':
          description: List of cinemas showing the movie
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    cinema:
                      type: string
                      example: "Cines Callao"
                    address:
                      type: string
                      example: "Plaza Callao 3 28013 Madrid"
                    showtimes:
                      type: object
                      additionalProperties:
                        type: array
                        items:
                          type: string
                          example: "16:00"
        '404':
          description: No cinemas found showing the specified movie
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "No cinemas found showing the specified movie"